outsourced App.js, renamed former App to MapInterface and setup components index
