+++ src/App.js
@@ -1,0 +1,18 @@
+import React, { Component } from "react";
+import Routes from './components/';
+import { BrowserRouter as Router } from "react-router-dom";
+
+import "./tachyons.css";
+
+
+export default class App extends Component {
+  render() {
+    return (
+      <Router>
+        <div className="App">
+          <Routes />
+        </div>
+      </Router>
+    )
+  }
+}
